+++ src/index.ts
@@ -42,6 +42,7 @@ import {
 } from "./types";
 
 const TIME_DECAY = WeightName.TIME_DECAY;
+const TRENDING = WeightName.TRENDING;
 
 
 class TheAlgorithm {
@@ -95,8 +96,11 @@ class TheAlgorithm {
             scorerInfos[scorer.name] = scorer.getInfo();
             return scorerInfos;
         },
-        // TimeDecay requires bespoke handling so it's not included in the loop above
-        {[TIME_DECAY]: Object.assign({}, DEFAULT_WEIGHTS[TIME_DECAY])} as ScorerDict
+        // TimeDecay and Trending require bespoke handling so it's not included in the loop above
+        {
+            [TIME_DECAY]: Object.assign({}, DEFAULT_WEIGHTS[TIME_DECAY]),
+            [TRENDING]: Object.assign({}, DEFAULT_WEIGHTS[TRENDING]),
+        } as ScorerDict
     );
 
     // This is the alternate constructor() that instantiates the class and loads the feed from storage.
@@ -402,6 +406,7 @@ class TheAlgorithm {
 
 export {
     TIME_DECAY,
+    TRENDING,
     FeedFilterSettings,
     NumericFilter,
     PropertyFilter,