+++ src/types.ts
@@ -39,11 +39,13 @@ export enum WeightName {
     NUM_REPLIES = 'NumReplies',
     NUM_RETOOTS = 'NumRetoots',
     RETOOTED_IN_FEED = 'RetootedInFeed',
-    TIME_DECAY = 'TimeDecay',
     TRENDING_LINKS = 'TrendingLinks',
     TRENDING_TAGS = "TrendingTags",
     TRENDING_TOOTS = "TrendingToots",
     VIDEO_ATTACHMENTS = 'VideoAttachments',
+    // Special weights
+    TIME_DECAY = 'TimeDecay',
+    TRENDING = 'Trending',
 };